refactored execute function from __main__.py to workspace.py, removed getansible.sh check in tests
